+++ upcloud/resource_upcloud_firewall_rules_test.go
@@ -97,10 +97,9 @@ func TestUpcloudFirewallRules_import(t *testing.T) {
 				),
 			},
 			{
-				ResourceName:            resourceName,
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"server_id"},
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})